src/test/java/models/SneakerTest.java
package models;

import junit.framework.TestCase;
import org.junit.Assert;
import org.junit.Test;

public class SneakerTest extends TestCase{
    @Test // (1)
    public void constructorTest(){
        // (2)
        int expectedId = 6;
        String expectedName = "Stan Smith";
        String expectedBrand = "Adidas";
        String expectedSport = "Tennnis";
        int expectedQty = 10;
        float expectedPrice = 80.00f;
        float expectedSize = 10.5f;

        // (3)
        Sneaker testSneaker = new Sneaker(expectedId, expectedName, expectedBrand,
                expectedSport, expectedSize,expectedQty,expectedPrice);

        // (4)
        Assert.assertEquals(expectedId, testSneaker.getId());
        Assert.assertEquals(expectedName, testSneaker.getName());
        Assert.assertEquals(expectedBrand, testSneaker.getBrand());
        Assert.assertEquals(expectedSport, testSneaker.getSport());
        Assert.assertEquals(expectedQty, testSneaker.getQty());
        Assert.assertEquals(java.util.Optional.of(expectedPrice), testSneaker.getPrice());
    }

}